migrated StatusLabel component to ts
