+++ src/Status/StatusLabel/StatusLabel.stories.tsx
@@ -1,13 +1,23 @@
+import { Meta, Story } from "@storybook/react";
+
 import React from "react";
 import StatusLabel from "./StatusLabel";
+import { StatusLabelProps } from "./StatusLabel.types";
 import { statusTypes } from "../statuses";
 
-export default {
+/**
+ * Story metadata
+ */
+const meta: Meta<typeof StatusLabel> = {
   component: StatusLabel,
   title: "Status/StatusLabel"
 };
+export default meta;
 
-const Template = args => {
+/**
+ * Story template for the StatusLabel component
+ */
+const Template: Story<StatusLabelProps> = args => {
   return <StatusLabel {...args} />;
 };
