+++ src/Status/StatusLabel/StatusLabel.test.tsx
@@ -3,7 +3,12 @@ import * as React from "react";
 import statuses, { statusTypes } from "../statuses";
 
 import StatusLabel from "./StatusLabel";
+import { SvgIconComponent } from "@mui/icons-material";
 import { render } from "@testing-library/react";
+
+interface CustomIcon extends SvgIconComponent {
+  name?: string;
+}
 
 describe("StatusLabel", () => {
   test.each(statusTypes)("renders correct icon %s", statusType => {
@@ -11,7 +16,7 @@ describe("StatusLabel", () => {
     // check icon is correct
     expect(container.querySelector("svg")).toHaveAttribute(
       "data-testid",
-      statuses[statusType].icon.type.name
+      (statuses[statusType].icon.type as CustomIcon).name
     );
   });
 });
