+++ src/Status/StatusLabel/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./StatusLabel";
+export { StatusLabelProps } from "./StatusLabel.types";
