+++ src/Status/index.ts
@@ -1,3 +1,3 @@
 export { default as StatusIcon, StatusIconProps } from "./StatusIcon";
-export { default as StatusLabel } from "./StatusLabel";
+export { default as StatusLabel, StatusLabelProps } from "./StatusLabel";
 export { default as StatusCard, StatusCardProps } from "./StatusCard";
